requirements.txt
accelerate
codetiming
datasets
flash-attn==2.7.4.post1
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
scipy
matplotlib
ipykernel
sam2
pycocotools
scikit-image
ray
tensordict
transformers==4.51.3
vllm==0.8.5.post1
wandb
